+++ NEWS.md
@@ -1,3 +1,27 @@
+## Roadmap 0.40
+
+- Intraday Event Study
+- Fixing date issue: dates before 01.01.1970 are currently not working
+
+
+## Version 0.33
+
+- add Excel export functionality to ResultParser
+- add new file ckecks
+    - check event window is in firm and market data
+    - check estimation window is in firm and market data
+    - dates before 01.01.1970 are actually not allowed
+- add function estAPIKey for setting EventStudy API key
+- ResultParser now parses CAR values
+
+### Bugfixes
+
+- fix EventStudy Addin 
+    - tradding days options
+    - statistics options
+    - object is now returned to console
+
+
 ## Version 0.32
 
 - redesign vignettes